hub-server/requirements.txt
# LeLamp Hub Server Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0
# clerk-sdk-python>=0.1.0  # Uncomment when using Clerk SDK

# File handling
python-multipart>=0.0.6
aiofiles>=23.0.0

# Google Cloud Storage (optional)
# google-cloud-storage>=2.0.0

# Utilities
httpx>=0.25.0
